+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix dropbar dropdown alignment if boundary-align is true
 - Fix viewport helper functions take border into account
 - Revert: Fix reposition Drop on scroll
+- Fix Toggle component does not `defaultPrevent()` click on anchor with touch gesture
 
 ## 3.13.9 (April 20, 2022)